separate auth components
